Completed the Search Window. Can serve as an example to MVVM with WPF Databindings

+++ InvoiceSystem/Domain/Item.cs
@@ -26,6 +26,6 @@ namespace InvoiceSystem
         /// <summary>
         /// The cost of an item
         /// </summary>
-        public float Cost { get; set; }
+        public decimal Cost { get; set; }
     }
 }
